pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weatherman"
version = "0.1.0"
description = "Unified time-series forecasting interface over Nixtla + AutoGluon"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Bryan Whiting" }]
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26",
  "pydantic>=2.8",
  "statsforecast>=1.7.8",
  "utilsforecast>=0.2.6",
]

[project.optional-dependencies]
autogluon = ["autogluon.timeseries>=1.2"]
dev = ["pytest>=8.0", "ruff>=0.6"]

[project.scripts]
weatherman = "weatherman.cli:main"

[tool.setuptools]
packages = ["weatherman"]

[tool.ruff]
line-length = 100
